egl: move loading into separate function
